Module embedded_time::duration

Duration types/units

Re-exports

pub use crate::fraction::Fraction;

Structs

Generic
The Generic Duration type allows an arbitrary scaling factor to be used without having to impl FixedPoint.

Hours
A duration unit type

Microseconds
A duration unit type

Milliseconds
A duration unit type

Minutes
A duration unit type

Nanoseconds
A duration unit type

Seconds
A duration unit type

Traits

Duration
An unsigned, fixed-point duration type

Extensions
Create duration-based extensions from primitive numeric types.