Struct embedded_time::duration::Generic

pub struct Generic<T> { /* fields omitted */ }

The Generic Duration type allows an arbitrary scaling factor to be used without having to impl FixedPoint.

The purpose of this type is to allow a simple Duration object that can be defined at run-time. It does this by replacing the const scaling factor with a struct field.

Implementations

impl<T> Generic<T>

pub const fn new(integer: T, scaling_factor: Fraction) -> Self

Constructs a new fixed-point Generic Duration value

pub const fn integer(&self) -> &T

Returns the integer part

pub const fn scaling_factor(&self) -> &Fraction

Returns the scaling factor Fraction part