Struct embedded_time::fraction::Fraction

pub struct Fraction(_);

A fractional value

Used primarily to define the scaling factor for the Duration, Rate, Instant and Clock traits and types.

Implementations

impl Fraction

pub const fn new(numerator: u32, denominator: u32) -> Self

Construct a new Fraction.

A reduction is not performed. Also there is no check for a denominator of 0. If these features are needed, use Fraction::new_reduce()

pub const fn numerator(&self) -> &u32

Return the numerator of the fraction

pub const fn denominator(&self) -> &u32

Return the denominator of the fraction

impl Fraction

pub fn new_reduce(
numerator: u32,
denominator: u32
) -> Result<Self, ConversionError>

Construct a new Fraction.

A reduction and denominator == 0 check are performed.

Errors

ConversionError::DivByZero : A 0 denominator was detected

pub fn to_integer(&self) -> u32

Returns the value truncated to an integer

pub fn from_integer(value: u32) -> Self

Constructs a Fraction from an integer.

Equivalent to Fraction::new(value,1).

pub fn recip(self) -> Self

Returns the reciprocal of the fraction

pub fn checked_mul(&self, v: &Self) -> Option<Self>

Checked Fraction × Fraction = Fraction

Returns None for any errors

Examples

assert_eq!(Fraction::new(1000, 1).checked_mul(&Fraction::new(5,5)),
    Some(Fraction::new(5_000, 5)));

assert_eq!(Fraction::new(u32::MAX, 1).checked_mul(&Fraction::new(2,1)),
    None);

pub fn checked_div(&self, v: &Self) -> Option<Self>

Checked Fraction / Fraction = Fraction

Returns None for any errors

Examples

assert_eq!(Fraction::new(1000, 1).checked_div(&Fraction::new(10, 1000)),
    Some(Fraction::new(1_000_000, 10)));

assert_eq!(Fraction::new(1, u32::MAX).checked_div(&Fraction::new(2,1)),
    None);