Module embedded_time::rate

Rate-based types/units

Re-exports

pub use crate::fraction::Fraction;

Structs

Baud
Baud

BitsPerSecond
Bits/s

BytesPerSecond
Bytes/s

Generic
The Generic Rate type allows an arbitrary scaling factor to be used without having to impl FixedPoint.

Hertz
Hertz

Kibibaud
Baud × 1,024

KibibitsPerSecond
Bits/s × 1,024

KibibytesPerSecond
Bytes/s × 1,024

Kibihertz
Hertz × 1,024

Kilobaud
Baud × 1,000

KilobitsPerSecond
Bits/s × 1,000

KilobytesPerSecond
Bytes/s × 1,000

Kilohertz
Hertz × 1,000

Mebibaud
Baud × 1,048,576

MebibitsPerSecond
Bits/s × 1,048,576

MebibytesPerSecond
Bytes/s × 1,048,576

Mebihertz
Hertz × 1,048,576

Megabaud
Baud × 1,000,000

MegabitsPerSecond
Bits/s × 1,000,000

MegabytesPerSecond
Bytes/s × 1,000,000

Megahertz
Hertz × 1,000,000

Traits

Extensions
Create rate-based extensions from primitive numeric types.

Rate
An unsigned, fixed-point rate type