Struct embedded_time::rate::Generic

pub struct Generic<T> { /* private fields */ }

The Generic Rate type allows an arbitrary scaling factor to be used without having to impl FixedPoint.

The purpose of this type is to allow a simple Rate object that can be defined at run-time. It does this by replacing the const scaling factor with a struct field.

Implementations

impl<T: TimeInt> Generic<T>

pub fn new(integer: T, scaling_factor: Fraction) -> Self

Constructs a new fixed-point Generic Rate value

pub fn integer(&self) -> T

Returns the integer part

pub fn scaling_factor(&self) -> &Fraction

Returns the scaling factor Fraction part