Struct embedded_time::timer::Timer

pub struct Timer<'a, Type, State, Clock: Clock, Dur: Duration> { /* private fields */ }

A Timer counts toward an expiration, can be polled for elapsed and remaining time, and can be one-shot or continuous/periodic.

Implementations

impl<'a, Clock: Clock, Dur: Duration> Timer<'_, None, None, Clock, Dur>

pub fn new(clock: &Clock, duration: Dur) -> Timer<'_, OneShot, Armed, Clock, Dur>

Construct a new, OneShot Timer

impl<'a, Type, State, Clock: Clock, Dur: Duration> Timer<'a, Type, State, Clock, Dur>

pub fn into_oneshot(self) -> Timer<'a, OneShot, State, Clock, Dur>

Change timer type to one-shot

pub fn into_periodic(self) -> Timer<'a, Periodic, State, Clock, Dur>

Change timer type into periodic

impl<'a, Type, Clock: Clock, Dur: Duration> Timer<'a, Type, Armed, Clock, Dur>

pub fn start(self) -> Result<Timer<'a, Type, Running, Clock, Dur>, TimeError>where
Clock::T: TryFrom<Dur::T>,
Dur: FixedPoint,

Start the timer from this instant

impl<Type, Clock: Clock, Dur: Duration> Timer<'_, Type, Running, Clock, Dur>

pub fn elapsed(&self) -> Result<Dur, TimeError>where
Dur: FixedPoint + TryFrom<Generic<Clock::T>, Error = ConversionError>,
Dur::T: TryFrom<Clock::T>,
Clock::T: TryFrom<Dur::T>,

Returns the Duration of time elapsed since it was started

The duration is truncated, not rounded.

The units of the Duration are the same as that used to construct the Timer.

pub fn remaining(&self) -> Result<Dur, TimeError>where
Dur: FixedPoint + TryFrom<Generic<Clock::T>, Error = ConversionError>,
Dur::T: TryFrom<u32> + TryFrom<Clock::T>,
Clock::T: TryFrom<Dur::T>,

Returns the Duration until the expiration of the timer

The duration is truncated, not rounded.

The units of the Duration are the same as that used to construct the Timer.

impl<'a, Clock: Clock, Dur: Duration> Timer<'a, OneShot, Running, Clock, Dur>

pub fn wait(self) -> Result<Timer<'a, OneShot, Armed, Clock, Dur>, TimeError>

Block until the timer has expired

pub fn is_expired(&self) -> Result<bool, TimeError>

Check whether the timer has expired

The timer is not restarted

impl<Clock: Clock, Dur: Duration> Timer<'_, Periodic, Running, Clock, Dur>

pub fn wait(self) -> Result<Self, TimeError>where
Instant<Clock>: Add<Dur, Output = Instant<Clock>>,

Block until the timer has expired

The timer is restarted

pub fn period_complete(&mut self) -> Result<bool, TimeError>where
Instant<Clock>: Add<Dur, Output = Instant<Clock>>,

Check whether a periodic timer has elapsed

The timer is restarted if it has elapsed.